move some tests back to libcollections
